Removed the 'prerelease' type and added conditional logic around updating param store
I was using the 'prerelease' type to debug values in the release event context. Now that we've got
the script sorted out, let's remove that type. We don't want to update the download URL on prerelease.

Also, it’s worth adding some conditional logic around those changes in the param store just in case we
don’t get a value from jq select (the filename we use for the desktop app has a convention and if we
upload a new version with a slightly different name, this would break)